requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.29
streamlit==1.32.2
PyMuPDF==1.24.0
chromadb==0.4.24
pyngrok==7.1.5
langchain-core>=0.1.10
openai>=1.3.0
pandas>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0  # for Excel file support